chore: extract onRecordJourney
