+++ src/App.tsx
@@ -50,6 +50,7 @@ import { ExportScriptFlyout } from './components/ExportScriptFlyout';
 import { useRecordingContext } from './hooks/useRecordingContext';
 import { StartOverWarningModal } from './components/StartOverWarningModal';
 import { ActionContext, RecorderSteps, Steps } from '../common/types';
+import type { IpcRendererEvent } from 'electron';
 
 /**
  * This is the prescribed workaround to some internal EUI issues that occur
@@ -84,16 +85,15 @@ export default function App() {
 
   useEffect(() => {
     // `actions` here is a set of `ActionInContext`s that make up a `Step`
-    const listener = ({ actions }: { actions: ActionContext[] }) => {
+    const listener = (_event: IpcRendererEvent, actions: ActionContext[]) => {
       setSteps((prevSteps: RecorderSteps) => {
         const nextSteps: Steps = generateIR([{ actions }]);
         return generateMergedIR(prevSteps, nextSteps);
       });
     };
-    ipc.answerMain('change', listener);
-    return () => {
-      ipc.removeListener('change', listener);
-    };
+    // ipc.answerMain('change', listener);
+    const removeListener = window.electronAPI.onActionGenerated(listener);
+    return removeListener;
   }, [ipc, setSteps]);
 
   return (
